Day 20 part 2
This was strange...

I ended up visualizing the input graph with graphviz, noticing that
there are four distinct but similar structures, each of them has to
output low value to get rx low.

So I checked for each of them when is the first time they output low,
ran lcm on this and this passed. I was prepared to have to properly
detects cycles for each of those structures, but apparently they are
created in such a way to cycle on outputing low.

No idea if this can be nicely solved generically ;)